Allow optional platform gateway tokens in bitloops-inference
