+++ bitloops/src/config/daemon_config.rs
@@ -1,1041 +1,24 @@
-use anyhow::{Context, Result, bail};
-use semver::Version;
-use serde::Deserialize;
-use serde_json::Value;
-use std::fs;
-use std::path::{Path, PathBuf};
-use toml_edit::{Array, DocumentMut, Item, Table, Value as TomlValue, de::from_str};
-
-use crate::host::inference::{
-    BITLOOPS_EMBEDDINGS_IPC_DRIVER, BITLOOPS_INFERENCE_RUNTIME_ID,
-    BITLOOPS_LOCAL_EMBEDDINGS_RUNTIME_ID, BITLOOPS_PLATFORM_EMBEDDINGS_RUNTIME_ID,
-};
-use crate::utils::platform_dirs::{bitloops_config_file_path, ensure_dir, ensure_parent_dir};
-
-use super::resolve_blob_local_path_for_repo;
-use super::unified_config::{UnifiedSettings, resolve_store_backend_from_unified};
-
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub struct DaemonCliSettings {
-    pub local_dev: bool,
-    pub cli_version: String,
-    pub telemetry: Option<bool>,
-    pub log_level: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct DaemonTelemetryConsentState {
-    pub telemetry: Option<bool>,
-    pub cli_version: String,
-    pub needs_prompt: bool,
-}
-
-#[derive(Debug, Clone)]
-pub struct LoadedDaemonSettings {
-    pub path: PathBuf,
-    pub root: PathBuf,
-    pub settings: UnifiedSettings,
-    pub cli: DaemonCliSettings,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(crate) enum DaemonEmbeddingsInstallMode {
-    Bootstrap,
-    WarmExisting,
-    SkipHosted,
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct DaemonEmbeddingsInstallPlan {
-    pub config_path: PathBuf,
-    pub profile_name: String,
-    pub runtime_name: String,
-    pub profile_driver: Option<String>,
-    pub mode: DaemonEmbeddingsInstallMode,
-    pub config_modified: bool,
-    original_contents: Option<String>,
-    prepared_contents: Option<String>,
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct DaemonInferenceInstallPlan {
-    pub config_path: PathBuf,
-    pub config_modified: bool,
-    original_contents: Option<String>,
-    prepared_contents: Option<String>,
-}
-
-impl DaemonEmbeddingsInstallPlan {
-    pub fn apply(&self) -> Result<()> {
-        self.write_prepared_contents(self.prepared_contents.as_deref())
-    }
-
-    pub fn apply_with_managed_runtime_path(&self, binary_path: &Path) -> Result<()> {
-        if !self.config_modified {
-            return Ok(());
-        }
-
-        let staged_contents = self.prepared_contents.as_deref().unwrap_or_default();
-        let mut staged_doc = staged_contents.parse::<DocumentMut>().with_context(|| {
-            format!(
-                "parsing staged Bitloops daemon config {}",
-                self.config_path.display()
-            )
-        })?;
-        let inference = ensure_table(&mut staged_doc, "inference");
-        let runtimes = ensure_child_table(inference, "runtimes");
-        let runtime = ensure_child_table(runtimes, &self.runtime_name);
-        runtime["command"] = Item::Value(binary_path.to_string_lossy().to_string().into());
-
-        let desired_runtime = staged_doc
-            .get("inference")
-            .and_then(Item::as_table)
-            .and_then(|table| table.get("runtimes"))
-            .and_then(Item::as_table)
-            .and_then(|table| table.get(&self.runtime_name))
-            .cloned()
-            .context("staged embeddings runtime missing from prepared config")?;
-        let desired_profile = staged_doc
-            .get("inference")
-            .and_then(Item::as_table)
-            .and_then(|table| table.get("profiles"))
-            .and_then(Item::as_table)
-            .and_then(|table| table.get(&self.profile_name))
-            .cloned()
-            .with_context(|| {
-                format!(
-                    "staged embeddings profile `{}` missing from prepared config",
-                    self.profile_name
-                )
-            })?;
-        let desired_code_embeddings = staged_doc
-            .get("semantic_clones")
-            .and_then(Item::as_table)
-            .and_then(|table| table.get("inference"))
-            .and_then(Item::as_table)
-            .and_then(|table| table.get("code_embeddings"))
-            .cloned()
-            .context("staged code_embeddings binding missing from prepared config")?;
-        let desired_summary_embeddings = staged_doc
-            .get("semantic_clones")
-            .and_then(Item::as_table)
-            .and_then(|table| table.get("inference"))
-            .and_then(Item::as_table)
-            .and_then(|table| table.get("summary_embeddings"))
-            .cloned()
-            .context("staged summary_embeddings binding missing from prepared config")?;
-
-        let current_contents = match fs::read_to_string(&self.config_path) {
-            Ok(contents) => contents,
-            Err(err) if err.kind() == std::io::ErrorKind::NotFound => String::new(),
-            Err(err) => {
-                return Err(err).with_context(|| {
-                    format!(
-                        "reading current Bitloops daemon config {}",
-                        self.config_path.display()
-                    )
-                });
-            }
-        };
-        let mut current_doc = if current_contents.trim().is_empty() {
-            DocumentMut::new()
-        } else {
-            current_contents.parse::<DocumentMut>().with_context(|| {
-                format!(
-                    "parsing current Bitloops daemon config {}",
-                    self.config_path.display()
-                )
-            })?
-        };
-
-        {
-            let inference = ensure_table(&mut current_doc, "inference");
-            let runtimes = ensure_child_table(inference, "runtimes");
-            runtimes[&self.runtime_name] = desired_runtime;
-            let profiles = ensure_child_table(inference, "profiles");
-            profiles[&self.profile_name] = desired_profile;
-        }
-
-        {
-            let semantic_clones = ensure_table(&mut current_doc, "semantic_clones");
-            let inference = ensure_child_table(semantic_clones, "inference");
-            inference["code_embeddings"] = desired_code_embeddings;
-            inference["summary_embeddings"] = desired_summary_embeddings;
-        }
-
-        self.write_prepared_contents(Some(&current_doc.to_string()))
-    }
-
-    pub fn rollback(&self) -> Result<()> {
-        if !self.config_modified {
-            return Ok(());
-        }
-
-        match &self.original_contents {
-            Some(contents) => fs::write(&self.config_path, contents).with_context(|| {
-                format!(
-                    "restoring Bitloops daemon config after failed embeddings install {}",
-                    self.config_path.display()
-                )
-            })?,
-            None => {
-                if self.config_path.exists() {
-                    fs::remove_file(&self.config_path).with_context(|| {
-                        format!(
-                            "removing Bitloops daemon config after failed embeddings install {}",
-                            self.config_path.display()
-                        )
-                    })?;
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    fn write_prepared_contents(&self, contents: Option<&str>) -> Result<()> {
-        if !self.config_modified {
-            return Ok(());
-        }
-
-        let Some(contents) = contents else {
-            return Ok(());
-        };
-
-        fs::write(&self.config_path, contents).with_context(|| {
-            format!(
-                "writing Bitloops daemon config {}",
-                self.config_path.display()
-            )
-        })
-    }
-}
-
-impl DaemonInferenceInstallPlan {
-    pub fn apply(&self) -> Result<()> {
-        self.write_prepared_contents(self.prepared_contents.as_deref())
-    }
-
-    pub fn rollback(&self) -> Result<()> {
-        if !self.config_modified {
-            return Ok(());
-        }
-
-        match &self.original_contents {
-            Some(contents) => fs::write(&self.config_path, contents).with_context(|| {
-                format!(
-                    "restoring Bitloops daemon config after failed inference install {}",
-                    self.config_path.display()
-                )
-            })?,
-            None => {
-                if self.config_path.exists() {
-                    fs::remove_file(&self.config_path).with_context(|| {
-                        format!(
-                            "removing Bitloops daemon config after failed inference install {}",
-                            self.config_path.display()
-                        )
-                    })?;
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    fn write_prepared_contents(&self, contents: Option<&str>) -> Result<()> {
-        if !self.config_modified {
-            return Ok(());
-        }
-
-        let Some(contents) = contents else {
-            return Ok(());
-        };
-
-        fs::write(&self.config_path, contents).with_context(|| {
-            format!(
-                "writing Bitloops daemon config {}",
-                self.config_path.display()
-            )
-        })
-    }
-}
-
-#[derive(Debug, Deserialize, Default)]
-#[serde(deny_unknown_fields)]
-struct DaemonTomlFile {
-    #[serde(default)]
-    runtime: RuntimeToml,
-    #[serde(default)]
-    telemetry: TelemetryToml,
-    #[serde(default)]
-    logging: LoggingToml,
-    #[serde(default)]
-    stores: Option<Value>,
-    #[serde(default)]
-    knowledge: Option<Value>,
-    #[serde(default)]
-    semantic_clones: Option<Value>,
-    #[serde(default)]
-    inference: Option<Value>,
-    #[serde(default)]
-    dashboard: Option<Value>,
-}
-
-#[derive(Debug, Deserialize, Default)]
-#[serde(deny_unknown_fields)]
-struct RuntimeToml {
-    #[serde(default)]
-    local_dev: bool,
-    cli_version: Option<String>,
-}
-
-#[derive(Debug, Deserialize, Default)]
-#[serde(deny_unknown_fields)]
-struct TelemetryToml {
-    enabled: Option<bool>,
-}
-
-#[derive(Debug, Deserialize, Default)]
-#[serde(deny_unknown_fields)]
-struct LoggingToml {
-    level: Option<String>,
-}
-
-pub fn default_daemon_config_path() -> Result<PathBuf> {
-    bitloops_config_file_path()
-}
-
-pub fn default_daemon_config_exists() -> Result<bool> {
-    Ok(default_daemon_config_path()?.is_file())
-}
-
-pub fn load_daemon_settings(explicit_path: Option<&Path>) -> Result<LoadedDaemonSettings> {
-    let path = match explicit_path {
-        Some(path) => path.to_path_buf(),
-        None => default_daemon_config_path()?,
-    };
-    let root = path
-        .parent()
-        .map(Path::to_path_buf)
-        .context("resolving Bitloops daemon config directory")?;
-
-    let file = match fs::read_to_string(&path) {
-        Ok(data) => parse_daemon_config_text(&data, &path)?,
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound && explicit_path.is_none() => {
-            bail!(
-                "Bitloops daemon config not found at {}. Run `bitloops start --create-default-config` or `bitloops init --install-default-daemon`.",
-                path.display()
-            );
-        }
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => {
-            bail!("Bitloops daemon config not found at {}", path.display());
-        }
-        Err(err) => {
-            return Err(err)
-                .with_context(|| format!("reading Bitloops daemon config {}", path.display()));
-        }
-    };
-
-    let cli = DaemonCliSettings {
-        local_dev: file.runtime.local_dev,
-        cli_version: file.runtime.cli_version.unwrap_or_default(),
-        telemetry: file.telemetry.enabled,
-        log_level: file.logging.level.unwrap_or_default(),
-    };
-
-    let canonical_path = path.canonicalize().unwrap_or_else(|_| path.clone());
-    let canonical_root = canonical_path
-        .parent()
-        .map(Path::to_path_buf)
-        .unwrap_or(root.clone());
-
-    Ok(LoadedDaemonSettings {
-        path: canonical_path,
-        root: canonical_root,
-        settings: UnifiedSettings {
-            enabled: None,
-            strategy: None,
-            local_dev: Some(cli.local_dev),
-            log_level: (!cli.log_level.is_empty()).then(|| cli.log_level.clone()),
-            strategy_options: None,
-            telemetry: cli.telemetry,
-            stores: file.stores,
-            knowledge: file.knowledge,
-            semantic_clones: file.semantic_clones,
-            inference: file.inference,
-            dashboard: file.dashboard,
-            watch: None,
-        },
-        cli,
-    })
-}
-
-fn parse_daemon_config_text(data: &str, path: &Path) -> Result<DaemonTomlFile> {
-    from_str::<DaemonTomlFile>(data)
-        .with_context(|| format!("parsing Bitloops daemon config {}", path.display()))
-}
-
-pub fn ensure_daemon_config_exists() -> Result<PathBuf> {
-    let path = default_daemon_config_path()?;
-    if path.exists() {
-        return Ok(path);
-    }
-
-    ensure_parent_dir(&path)?;
-    if let Some(parent) = path.parent() {
-        ensure_dir(parent)?;
-    }
-    fs::write(&path, default_daemon_config_toml()?)
-        .with_context(|| format!("writing Bitloops daemon config {}", path.display()))?;
-    Ok(path)
-}
-
-pub fn bootstrap_default_daemon_environment() -> Result<PathBuf> {
-    let path = ensure_daemon_config_exists()?;
-    ensure_daemon_store_artifacts(Some(path.as_path()))?;
-    Ok(path)
-}
-
-pub fn ensure_daemon_store_artifacts(explicit_path: Option<&Path>) -> Result<PathBuf> {
-    let loaded = load_daemon_settings(explicit_path)?;
-    ensure_local_store_artifacts(&loaded)?;
-    Ok(loaded.path)
-}
-
-pub fn persist_daemon_cli_settings(update: &DaemonCliSettings) -> Result<PathBuf> {
-    persist_daemon_cli_settings_at(None, update)
-}
-
-pub fn update_daemon_telemetry_consent(
-    explicit_path: Option<&Path>,
-    current_cli_version: &str,
-    telemetry_override: Option<bool>,
-) -> Result<DaemonTelemetryConsentState> {
-    let loaded = load_daemon_settings(explicit_path)?;
-    let current = normalise_cli_version(current_cli_version)?;
-    let mut cli = loaded.cli;
-
-    if let Some(choice) = telemetry_override {
-        cli.telemetry = Some(choice);
-    } else if cli.telemetry == Some(false)
-        && should_clear_telemetry_for_version(cli.cli_version.as_str(), &current)
-    {
-        cli.telemetry = None;
-    }
-
-    cli.cli_version = current;
-    persist_daemon_cli_settings_at(Some(loaded.path.as_path()), &cli)?;
-
-    Ok(DaemonTelemetryConsentState {
-        telemetry: cli.telemetry,
-        cli_version: cli.cli_version,
-        needs_prompt: cli.telemetry.is_none(),
-    })
-}
-
-fn persist_daemon_cli_settings_at(
-    explicit_path: Option<&Path>,
-    update: &DaemonCliSettings,
-) -> Result<PathBuf> {
-    let path = default_daemon_config_path()?;
-    let path = explicit_path.map(Path::to_path_buf).unwrap_or(path);
-    ensure_parent_dir(&path)?;
-
-    let mut doc = match fs::read_to_string(&path) {
-        Ok(existing) => existing
-            .parse::<DocumentMut>()
-            .with_context(|| format!("parsing Bitloops daemon config {}", path.display()))?,
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => DocumentMut::new(),
-        Err(err) => {
-            return Err(err)
-                .with_context(|| format!("reading Bitloops daemon config {}", path.display()));
-        }
-    };
-
-    {
-        let runtime = ensure_table(&mut doc, "runtime");
-        runtime["local_dev"] = Item::Value(update.local_dev.into());
-        if update.cli_version.trim().is_empty() {
-            runtime.remove("cli_version");
-        } else {
-            runtime["cli_version"] = Item::Value(update.cli_version.clone().into());
-        }
-    }
-
-    {
-        let logging = ensure_table(&mut doc, "logging");
-        if update.log_level.trim().is_empty() {
-            logging.remove("level");
-        } else {
-            logging["level"] = Item::Value(update.log_level.clone().into());
-        }
-    }
-
-    {
-        let telemetry = ensure_table(&mut doc, "telemetry");
-        match update.telemetry {
-            Some(choice) => telemetry["enabled"] = Item::Value(choice.into()),
-            None => {
-                telemetry.remove("enabled");
-            }
-        }
-    }
-
-    fs::write(&path, doc.to_string())
-        .with_context(|| format!("writing Bitloops daemon config {}", path.display()))?;
-    Ok(path)
-}
-
-pub fn persist_dashboard_tls_hint(enabled: bool) -> Result<PathBuf> {
-    let path = default_daemon_config_path()?;
-    ensure_parent_dir(&path)?;
-
-    let mut doc = match fs::read_to_string(&path) {
-        Ok(existing) => existing
-            .parse::<DocumentMut>()
-            .with_context(|| format!("parsing Bitloops daemon config {}", path.display()))?,
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => DocumentMut::new(),
-        Err(err) => {
-            return Err(err)
-                .with_context(|| format!("reading Bitloops daemon config {}", path.display()));
-        }
-    };
-
-    let dashboard = ensure_table(&mut doc, "dashboard");
-    let local_dashboard = ensure_child_table(dashboard, "local_dashboard");
-    local_dashboard["tls"] = Item::Value(enabled.into());
-
-    fs::write(&path, doc.to_string())
-        .with_context(|| format!("writing Bitloops daemon config {}", path.display()))?;
-    Ok(path)
-}
-
-pub(crate) fn prepare_daemon_embeddings_install(
-    config_path: &Path,
-) -> Result<DaemonEmbeddingsInstallPlan> {
-    const DEFAULT_LOCAL_PROFILE: &str = "local_code";
-    const DEFAULT_LOCAL_MODEL: &str = "bge-m3";
-
-    ensure_parent_dir(config_path)?;
-
-    let original_contents = match fs::read_to_string(config_path) {
-        Ok(contents) => Some(contents),
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => None,
-        Err(err) => {
-            return Err(err).with_context(|| {
-                format!("reading Bitloops daemon config {}", config_path.display())
-            });
-        }
-    };
-
-    let mut doc = match original_contents.as_deref() {
-        Some(existing) => existing
-            .parse::<DocumentMut>()
-            .with_context(|| format!("parsing Bitloops daemon config {}", config_path.display()))?,
-        None => DocumentMut::new(),
-    };
-
-    if let Some(profile_name) = selected_inference_profile_name(&doc) {
-        let profile_driver = inference_driver_for_profile(&doc, &profile_name);
-        let mode = if profile_driver.as_deref() == Some(BITLOOPS_EMBEDDINGS_IPC_DRIVER) {
-            DaemonEmbeddingsInstallMode::WarmExisting
-        } else {
-            DaemonEmbeddingsInstallMode::SkipHosted
-        };
-        return Ok(DaemonEmbeddingsInstallPlan {
-            config_path: config_path.to_path_buf(),
-            profile_name,
-            runtime_name: BITLOOPS_LOCAL_EMBEDDINGS_RUNTIME_ID.to_string(),
-            profile_driver,
-            mode,
-            config_modified: false,
-            original_contents,
-            prepared_contents: None,
-        });
-    }
-
-    if let Some(kind) = inference_driver_for_profile(&doc, DEFAULT_LOCAL_PROFILE)
-        && kind != BITLOOPS_EMBEDDINGS_IPC_DRIVER
-    {
-        bail!(
-            "cannot install default local embeddings because profile `{DEFAULT_LOCAL_PROFILE}` already exists with driver `{kind}`"
-        );
-    }
-
-    let mut modified = false;
-    {
-        let semantic_clones = ensure_table(&mut doc, "semantic_clones");
-        let inference = ensure_child_table(semantic_clones, "inference");
-        if inference
-            .get("code_embeddings")
-            .and_then(Item::as_value)
-            .and_then(|value| value.as_str())
-            != Some(DEFAULT_LOCAL_PROFILE)
-        {
-            inference["code_embeddings"] = Item::Value(DEFAULT_LOCAL_PROFILE.into());
-            modified = true;
-        }
-        if inference
-            .get("summary_embeddings")
-            .and_then(Item::as_value)
-            .and_then(|value| value.as_str())
-            != Some(DEFAULT_LOCAL_PROFILE)
-        {
-            inference["summary_embeddings"] = Item::Value(DEFAULT_LOCAL_PROFILE.into());
-            modified = true;
-        }
-    }
-
-    {
-        let inference = ensure_table(&mut doc, "inference");
-
-        let runtimes = ensure_child_table(inference, "runtimes");
-        let runtime = ensure_child_table(runtimes, BITLOOPS_LOCAL_EMBEDDINGS_RUNTIME_ID);
-        let current_runtime_command = runtime
-            .get("command")
-            .and_then(Item::as_value)
-            .and_then(|value| value.as_str())
-            .map(str::trim)
-            .map(ToOwned::to_owned);
-        if runtime
-            .get("command")
-            .and_then(Item::as_value)
-            .and_then(|value| value.as_str())
-            .is_none()
-        {
-            runtime["command"] = Item::Value("bitloops-local-embeddings".into());
-            modified = true;
-        }
-
-        let manages_default_args = current_runtime_command.is_none()
-            || matches!(
-                current_runtime_command.as_deref(),
-                Some("")
-                    | Some("bitloops-local-embeddings")
-                    | Some("bitloops-local-embeddings.exe")
-            );
-        let runtime_args_are_empty = runtime
-            .get("args")
-            .and_then(Item::as_value)
-            .and_then(|value| value.as_array())
-            .is_some_and(|value| value.is_empty());
-        if (manages_default_args && !runtime_args_are_empty)
-            || !runtime.get("args").is_some_and(Item::is_value)
-        {
-            runtime["args"] = Item::Value(TomlValue::Array(Array::new()));
-            modified = true;
-        }
-        if runtime
-            .get("startup_timeout_secs")
-            .and_then(Item::as_value)
-            .and_then(|value| value.as_integer())
-            .is_none()
-        {
-            runtime["startup_timeout_secs"] = Item::Value(60.into());
-            modified = true;
-        }
-        if runtime
-            .get("request_timeout_secs")
-            .and_then(Item::as_value)
-            .and_then(|value| value.as_integer())
-            .is_none()
-        {
-            runtime["request_timeout_secs"] = Item::Value(300.into());
-            modified = true;
-        }
-
-        let profiles = ensure_child_table(inference, "profiles");
-        let local_profile = ensure_child_table(profiles, DEFAULT_LOCAL_PROFILE);
-        if local_profile
-            .get("task")
-            .and_then(Item::as_value)
-            .and_then(|value| value.as_str())
-            .is_none()
-        {
-            local_profile["task"] = Item::Value("embeddings".into());
-            modified = true;
-        }
-        if local_profile
-            .get("driver")
-            .and_then(Item::as_value)
-            .and_then(|value| value.as_str())
-            .is_none()
-        {
-            local_profile["driver"] = Item::Value(BITLOOPS_EMBEDDINGS_IPC_DRIVER.into());
-            modified = true;
-        }
-        if local_profile
-            .get("runtime")
-            .and_then(Item::as_value)
-            .and_then(|value| value.as_str())
-            .is_none()
-        {
-            local_profile["runtime"] = Item::Value(BITLOOPS_LOCAL_EMBEDDINGS_RUNTIME_ID.into());
-            modified = true;
-        }
-        if local_profile
-            .get("model")
-            .and_then(Item::as_value)
-            .and_then(|value| value.as_str())
-            .is_none()
-        {
-            local_profile["model"] = Item::Value(DEFAULT_LOCAL_MODEL.into());
-            modified = true;
-        }
-    }
-
-    let prepared_contents = modified.then(|| doc.to_string());
-
-    Ok(DaemonEmbeddingsInstallPlan {
-        config_path: config_path.to_path_buf(),
-        profile_name: DEFAULT_LOCAL_PROFILE.to_string(),
-        runtime_name: BITLOOPS_LOCAL_EMBEDDINGS_RUNTIME_ID.to_string(),
-        profile_driver: Some(BITLOOPS_EMBEDDINGS_IPC_DRIVER.to_string()),
-        mode: DaemonEmbeddingsInstallMode::Bootstrap,
-        config_modified: modified,
-        original_contents,
-        prepared_contents,
-    })
-}
-
-pub(crate) fn prepare_daemon_platform_embeddings_install(
-    config_path: &Path,
-    gateway_url: &str,
-    api_key_env: &str,
-) -> Result<DaemonEmbeddingsInstallPlan> {
-    const DEFAULT_PLATFORM_PROFILE: &str = "platform_code";
-    const DEFAULT_PLATFORM_MODEL: &str = "bge-m3";
-
-    ensure_parent_dir(config_path)?;
-
-    let original_contents = match fs::read_to_string(config_path) {
-        Ok(contents) => Some(contents),
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => None,
-        Err(err) => {
-            return Err(err).with_context(|| {
-                format!("reading Bitloops daemon config {}", config_path.display())
-            });
-        }
-    };
-
-    let mut doc = match original_contents.as_deref() {
-        Some(existing) => existing
-            .parse::<DocumentMut>()
-            .with_context(|| format!("parsing Bitloops daemon config {}", config_path.display()))?,
-        None => DocumentMut::new(),
-    };
-
-    {
-        let semantic_clones = ensure_table(&mut doc, "semantic_clones");
-        let inference = ensure_child_table(semantic_clones, "inference");
-        if inference
-            .get("code_embeddings")
-            .and_then(Item::as_value)
-            .and_then(|value| value.as_str())
-            != Some(DEFAULT_PLATFORM_PROFILE)
-        {
-            inference["code_embeddings"] = Item::Value(DEFAULT_PLATFORM_PROFILE.into());
-        }
-        if inference
-            .get("summary_embeddings")
-            .and_then(Item::as_value)
-            .and_then(|value| value.as_str())
-            != Some(DEFAULT_PLATFORM_PROFILE)
-        {
-            inference["summary_embeddings"] = Item::Value(DEFAULT_PLATFORM_PROFILE.into());
-        }
-    }
-
-    {
-        let inference = ensure_table(&mut doc, "inference");
-        let runtimes = ensure_child_table(inference, "runtimes");
-        let runtime = ensure_child_table(runtimes, BITLOOPS_PLATFORM_EMBEDDINGS_RUNTIME_ID);
-        runtime["command"] = Item::Value("bitloops-platform-embeddings".into());
-
-        let mut args = Array::new();
-        args.push("--gateway-url");
-        args.push(gateway_url);
-        args.push("--api-key-env");
-        args.push(api_key_env);
-        runtime["args"] = Item::Value(TomlValue::Array(args));
-        runtime["startup_timeout_secs"] = Item::Value(60.into());
-        runtime["request_timeout_secs"] = Item::Value(300.into());
-
-        let profiles = ensure_child_table(inference, "profiles");
-        let platform_profile = ensure_child_table(profiles, DEFAULT_PLATFORM_PROFILE);
-        platform_profile["task"] = Item::Value("embeddings".into());
-        platform_profile["driver"] = Item::Value(BITLOOPS_EMBEDDINGS_IPC_DRIVER.into());
-        platform_profile["runtime"] = Item::Value(BITLOOPS_PLATFORM_EMBEDDINGS_RUNTIME_ID.into());
-        platform_profile["model"] = Item::Value(DEFAULT_PLATFORM_MODEL.into());
-    }
-
-    let prepared_contents = doc.to_string();
-    let config_modified = original_contents
-        .as_deref()
-        .is_none_or(|existing| existing != prepared_contents);
-
-    Ok(DaemonEmbeddingsInstallPlan {
-        config_path: config_path.to_path_buf(),
-        profile_name: DEFAULT_PLATFORM_PROFILE.to_string(),
-        runtime_name: BITLOOPS_PLATFORM_EMBEDDINGS_RUNTIME_ID.to_string(),
-        profile_driver: Some(BITLOOPS_EMBEDDINGS_IPC_DRIVER.to_string()),
-        mode: DaemonEmbeddingsInstallMode::Bootstrap,
-        config_modified,
-        original_contents,
-        prepared_contents: config_modified.then_some(prepared_contents),
-    })
-}
-
-pub(crate) fn prepare_daemon_inference_install(
-    config_path: &Path,
-) -> Result<DaemonInferenceInstallPlan> {
-    ensure_parent_dir(config_path)?;
-
-    let original_contents = match fs::read_to_string(config_path) {
-        Ok(contents) => Some(contents),
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => None,
-        Err(err) => {
-            return Err(err).with_context(|| {
-                format!("reading Bitloops daemon config {}", config_path.display())
-            });
-        }
-    };
-
-    let mut doc = match original_contents.as_deref() {
-        Some(existing) => existing
-            .parse::<DocumentMut>()
-            .with_context(|| format!("parsing Bitloops daemon config {}", config_path.display()))?,
-        None => DocumentMut::new(),
-    };
-
-    let mut modified = false;
-    let inference = ensure_table(&mut doc, "inference");
-    let runtimes = ensure_child_table(inference, "runtimes");
-    let runtime = ensure_child_table(runtimes, BITLOOPS_INFERENCE_RUNTIME_ID);
-    let current_runtime_command = runtime
-        .get("command")
-        .and_then(Item::as_value)
-        .and_then(|value| value.as_str())
-        .map(str::trim)
-        .map(ToOwned::to_owned);
-
-    if runtime
-        .get("command")
-        .and_then(Item::as_value)
-        .and_then(|value| value.as_str())
-        .is_none()
-    {
-        runtime["command"] = Item::Value("bitloops-inference".into());
-        modified = true;
-    }
-
-    let manages_default_args = current_runtime_command.is_none()
-        || matches!(
-            current_runtime_command.as_deref(),
-            Some("") | Some("bitloops-inference") | Some("bitloops-inference.exe")
-        );
-    let runtime_args_are_empty = runtime
-        .get("args")
-        .and_then(Item::as_value)
-        .and_then(|value| value.as_array())
-        .is_some_and(|value| value.is_empty());
-    if (manages_default_args && !runtime_args_are_empty)
-        || !runtime.get("args").is_some_and(Item::is_value)
-    {
-        runtime["args"] = Item::Value(TomlValue::Array(Array::new()));
-        modified = true;
-    }
-    if runtime
-        .get("startup_timeout_secs")
-        .and_then(Item::as_value)
-        .and_then(|value| value.as_integer())
-        .is_none()
-    {
-        runtime["startup_timeout_secs"] = Item::Value(60.into());
-        modified = true;
-    }
-    if runtime
-        .get("request_timeout_secs")
-        .and_then(Item::as_value)
-        .and_then(|value| value.as_integer())
-        .is_none()
-    {
-        runtime["request_timeout_secs"] = Item::Value(300.into());
-        modified = true;
-    }
-
-    let prepared_contents = modified.then(|| doc.to_string());
-
-    Ok(DaemonInferenceInstallPlan {
-        config_path: config_path.to_path_buf(),
-        config_modified: modified,
-        original_contents,
-        prepared_contents,
-    })
-}
-
-fn default_daemon_config_toml() -> Result<String> {
-    let mut doc = DocumentMut::new();
-    doc["runtime"] = Item::Table(Table::new());
-    doc["runtime"]["local_dev"] = Item::Value(false.into());
-
-    let default_root = Path::new(".");
-    let sqlite_path = crate::utils::paths::default_relational_db_path(default_root);
-    let duckdb_path = crate::utils::paths::default_events_db_path(default_root);
-    let blob_path = crate::utils::paths::default_blob_store_path(default_root);
-
-    doc["stores"] = Item::Table(Table::new());
-    doc["stores"]["relational"] = Item::Table(Table::new());
-    doc["stores"]["relational"]["sqlite_path"] =
-        Item::Value(sqlite_path.to_string_lossy().to_string().into());
-    doc["stores"]["events"] = Item::Table(Table::new());
-    doc["stores"]["events"]["duckdb_path"] =
-        Item::Value(duckdb_path.to_string_lossy().to_string().into());
-    doc["stores"]["blob"] = Item::Table(Table::new());
-    doc["stores"]["blob"]["local_path"] =
-        Item::Value(blob_path.to_string_lossy().to_string().into());
-
-    Ok(doc.to_string())
-}
-
-fn normalise_cli_version(current_cli_version: &str) -> Result<String> {
-    let trimmed = current_cli_version.trim();
-    if trimmed.is_empty() {
-        bail!("current CLI version must not be empty");
-    }
-    Version::parse(trimmed).context("current CLI version must be valid semver")?;
-    Ok(trimmed.to_string())
-}
-
-fn should_clear_telemetry_for_version(stored_cli_version: &str, current: &str) -> bool {
-    let trimmed = stored_cli_version.trim();
-    if trimmed.is_empty() {
-        return true;
-    }
-
-    let Ok(stored) = Version::parse(trimmed) else {
-        return true;
-    };
-    let Ok(current) = Version::parse(current) else {
-        return false;
-    };
-    stored < current
-}
-
-fn ensure_table<'a>(doc: &'a mut DocumentMut, key: &str) -> &'a mut Table {
-    let root = doc.as_table_mut();
-    if !root.contains_key(key) || !root[key].is_table() {
-        root.insert(key, Item::Table(Table::new()));
-    }
-    root[key]
-        .as_table_mut()
-        .expect("TOML item should be a table after initialisation")
-}
-
-fn ensure_child_table<'a>(table: &'a mut Table, key: &str) -> &'a mut Table {
-    if !table.contains_key(key) || !table[key].is_table() {
-        table.insert(key, Item::Table(Table::new()));
-    }
-    table[key]
-        .as_table_mut()
-        .expect("TOML item should be a table after initialisation")
-}
-
-fn selected_inference_profile_name(doc: &DocumentMut) -> Option<String> {
-    let inference = doc
-        .as_table()
-        .get("semantic_clones")?
-        .as_table()?
-        .get("inference")?
-        .as_table()?;
-
-    for key in ["code_embeddings", "summary_embeddings"] {
-        let Some(value) = inference
-            .get(key)
-            .and_then(Item::as_value)
-            .and_then(|value| value.as_str())
-            .map(str::trim)
-        else {
-            continue;
-        };
-        if value.is_empty() {
-            continue;
-        }
-        if matches!(
-            value.to_ascii_lowercase().as_str(),
-            "none" | "disabled" | "off"
-        ) {
-            continue;
-        }
-        return Some(value.to_string());
-    }
-
-    None
-}
-
-fn inference_driver_for_profile(doc: &DocumentMut, profile_name: &str) -> Option<String> {
-    doc.as_table()
-        .get("inference")?
-        .as_table()?
-        .get("profiles")?
-        .as_table()?
-        .get(profile_name)?
-        .as_table()?
-        .get("driver")?
-        .as_value()?
-        .as_str()
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(ToOwned::to_owned)
-}
-
-fn ensure_local_store_artifacts(loaded: &LoadedDaemonSettings) -> Result<()> {
-    let backends = resolve_store_backend_from_unified(&loaded.settings, &loaded.root)
-        .with_context(|| format!("resolving store backends from {}", loaded.path.display()))?;
-
-    if !backends.relational.has_postgres() {
-        let sqlite_path = backends
-            .relational
-            .resolve_sqlite_db_path_for_repo(&loaded.root)
-            .context("resolving SQLite path for daemon bootstrap")?;
-        if let Some(parent) = sqlite_path
-            .parent()
-            .filter(|parent| !parent.as_os_str().is_empty())
-        {
-            fs::create_dir_all(parent)
-                .with_context(|| format!("creating SQLite directory {}", parent.display()))?;
-        }
-        let _ = rusqlite::Connection::open(&sqlite_path)
-            .with_context(|| format!("creating SQLite database at {}", sqlite_path.display()))?;
-    }
-
-    if !backends.events.has_clickhouse() {
-        let duckdb_path = backends
-            .events
-            .resolve_duckdb_db_path_for_repo(&loaded.root);
-        if let Some(parent) = duckdb_path
-            .parent()
-            .filter(|parent| !parent.as_os_str().is_empty())
-        {
-            fs::create_dir_all(parent)
-                .with_context(|| format!("creating DuckDB directory {}", parent.display()))?;
-        }
-        let _ = duckdb::Connection::open(&duckdb_path)
-            .with_context(|| format!("creating DuckDB database at {}", duckdb_path.display()))?;
-    }
-
-    if !backends.blobs.has_remote() {
-        let blob_root =
-            resolve_blob_local_path_for_repo(&loaded.root, backends.blobs.local_path.as_deref())
-                .context("resolving blob store path for daemon bootstrap")?;
-        fs::create_dir_all(&blob_root)
-            .with_context(|| format!("creating blob store root {}", blob_root.display()))?;
-    }
-
-    Ok(())
-}
+#[path = "daemon_config/file.rs"]
+mod file;
+#[path = "daemon_config/install.rs"]
+mod install;
+#[path = "daemon_config/plans.rs"]
+mod plans;
+#[path = "daemon_config/toml.rs"]
+mod toml;
 
 #[cfg(test)]
 #[path = "daemon_config/tests.rs"]
 mod tests;
+
+pub use file::{
+    DaemonCliSettings, DaemonTelemetryConsentState, LoadedDaemonSettings,
+    bootstrap_default_daemon_environment, default_daemon_config_exists, default_daemon_config_path,
+    ensure_daemon_config_exists, ensure_daemon_store_artifacts, load_daemon_settings,
+    persist_daemon_cli_settings, persist_dashboard_tls_hint, update_daemon_telemetry_consent,
+};
+pub(crate) use install::{
+    prepare_daemon_embeddings_install, prepare_daemon_inference_install,
+    prepare_daemon_platform_embeddings_install,
+};
+pub(crate) use plans::DaemonEmbeddingsInstallMode;
